Expose replace method.
To replace a string with environment variables.

Substitution is performed on $variable and ${variable}.

If the referenced variable does not exist in the environment it is not
replaced, if the dollar is escaped with a backslash the replacement is
ignored regardless of whether the variable exists and the backslash
before the dollar is removed.